Enhance datamodule setup with new sequence parameters
Additional sequence parameters have been added to the data module setup. These parameters: `seq_len`, `out_seq_len`, `stride`, `batch_size`, and `num_workers` were added to enhance the flexibility and efficiency of the data handling process for PhyLSTM. This allows to fine-tune the model for different datasets and use-cases. The test cases for the enhanced data module were also correspondingly updated.

This further involved updating transformerft to transformertf in the updated files which incorporates the project-wide shift in nomenclature.