Fix reshaping issue in dataset operations
This commit is to address the reshaping bug identified in both 'x' and 'y' variables for instances where ndim equals 1. To ensure the right dimensionality for subsequent operations, both variables are now reshaped to (-1, 1) when they have ndim of 1. Tests were also updated to validate the use of derivatives (der) in both training and validation datasets. These tests now verify if the shapes of input and target data are as expected given a sequence length and possible multiple target values (FIELD, "der").