Add new fixture to load parquet data and refactor tests
A new pytest fixture was introduced in 'conftest.py' which loads a DataFrame from a parquet file, drops any null values, resets the index, and adds a new column for the derivative of a field. This fixture was used to refactor tests in 'test_phylstm_datamodule.py' and 'test_dataset.py', removing repetitive file loading and preprocessing steps. It improves the readability and maintainability of our testing code, allowing for easier adjustments in the future if necessary. The shape of the input data in 'test_dataset.py' was also updated to reflect recent changes in the data representation.