Update data transformation tests for PhyLSTM module and fix normalization bug
In the current diff, new tests have been added for checking the functionality of data transformation and inverse transformation for the PhyLSTMDataModule. This was needed to verify the correct operations of the transformation methods for the physical data module.

Also, some bugs were fixed in the `_normalizer.py` module where we ensured that the dimensions of the input scale parameters are compatible with the dimensions of the tensor being transformed. This step was necessary to avoid inconsistencies that could lead to failures when data of different dimensions are processed using the same normalizer instance.