Add target_scale to loss functions and adapt gradient logging
The commit adds a new 'target_scale' parameter to the loss functions in `_loss.py`. This facilitates loss calculations that need to take into account the scale of the target values - such as when using methods that scale data in the process. This adjustment broadens the flexibility and accuracy of these calculations. Furthermore, a boolean `log_grad_norm` was added to the model configuration to offer an option to log the gradient norm during training, which can assist in debugging and model optimization. Changes were also done to the dataset component to include 'target_scale' where necessary.