Change result assignment to new DataFrame in _normalize method
The current _normalize method in the TransformerTF data module applied transformations directly to columns of the input DataFrame. This has been changed to create a new DataFrame and apply transformations to it. This approach allows for keeping original input data intact and reduces potentials risks of inadvertent data modifications or loss. Same logic applied to _try_fit_polynomial_transform method.